requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
bidict==0.22.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.3
pycodestyle==2.10.0
pydantic==1.10.2
python-engineio==4.3.4
python-socketio==5.7.2
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
websockets==10.4
yarl==1.8.2